Design improvements from Hefty
increased bayonet insert slot length because on some printers, the fitment was too close.  Added retaining grove to base and producing rib around inside of cap.  Segmented cap to help the rib snap into grove.  Increased filter area slightly so that the filter itself is 80mm, not the outside of the filter base.  My mistake.